Fixed security, added filters and posts

+++ src/main/java/com/example/officeappbackend/dto/FilterDto.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 @Data
 public class FilterDto {
-    private List<Integer> officesId;
+    private List<Long> officesId;
     private Integer sortingFilterId;
 }
